security(deps): clear all RustSec advisories and gate on cargo-audit (#595)
Closes #584.

Clears every RustSec advisory in the lockfile, then adds the gate that
stops
them accumulating again. `cargo audit` now reports 0 vulnerabilities and
0
warnings across 563 crates, down from 11 vulnerabilities and 7 warnings
across
587.

## Remediation

Most of it was `cargo update` — the advisories were overwhelmingly
transitive:

| Crate | Was | Now | Advisories |
|---|---|---|---|
| crossbeam-epoch | 0.9.18 | 0.9.20 | RUSTSEC-2026-0204 |
| postgres-protocol | 0.6.11 | 0.6.12 | RUSTSEC-2026-0179, -0180 |
| tokio-postgres | 0.7.17 | 0.7.18 | RUSTSEC-2026-0178 |
| rustls-webpki | 0.103.10 | 0.103.14 | RUSTSEC-2026-0104, -0098, -0099
|
| quinn-proto | 0.11.14 | 0.11.17 | RUSTSEC-2026-0185 |

Two needed a manifest change:

- **protobuf 2.28.0 (RUSTSEC-2024-0437)** — reached only through
  `prometheus`'s default `protobuf` feature. `admin.rs` uses nothing but
`TextEncoder`, so `prometheus = { default-features = false }` drops the
dependency outright rather than waiting for prometheus to move to
protobuf
  3.x.
- **quick-xml 0.37.5 (RUSTSEC-2026-0194, -0195)** — our three direct
  dependants went to `0.41`, but the vulnerable copy survived under
`object_store 0.11.2`. `object_store 0.14.1` is the first release built
on
`quick-xml >= 0.41`; it is also the version `icechunk` already pulls
under
the `engine-zarr/icechunk` feature, so this collapses what were two
parallel
  object_store trees into one.

The two remaining informational warnings are not fixable from this repo
and are
documented with reachability analysis in `.cargo/audit.toml`, tracked in
#594:
`paste` (unmaintained, proc-macro only, absent from the runtime binary)
and
`lru` (unsound `pop()` panic path, reached only with POD cache payloads
that
cannot panic; no patched release exists).

## The quick-xml bump is behavioural, not cosmetic

Since 0.38 the reader no longer folds entity references into the
surrounding
text event — `&amp;` terminates the `Event::Text` and arrives as its own
`Event::GeneralRef`. Both CAP readers matched only `Event::Text`, so on
the
bumped version they would have **silently dropped every `&`, `<` and
`>`** from
alert headlines, descriptions and feed URLs. Silently, because the
document is
still well-formed and nothing errors — a `<link>` with a query string
would
just start 404ing and the collection would look empty.

`parser::push_general_ref` handles the new event (numeric char refs
resolved,
the five XML predefined entities expanded, anything else passed through
as
`&name;` rather than deleted). The reader-side `trim_text(true)` had to
go with
it: per-event trimming now eats the spaces *around* an entity, turning
`Wind &amp; Flood` into `Wind&Flood`. Every leaf was already trimmed
whole at
its closing tag, which is where trimming belongs. Four regression tests
pin all
of this.

`object_store 0.14` moved `get`/`head`/`get_range` from the
`ObjectStore` trait
into an `ObjectStoreExt` extension trait, and `get_range` now takes
`Range<u64>`. `DataStore` keeps its `Range<usize>` signature so no
caller
changes; the widening happens once at the boundary.

## The gate

A new `audit` job in `ci.yml`, running on PRs, on `main`, **and on the
weekly
schedule** — the schedule is the point: an advisory published against a
crate we
already ship turns CI red with nobody touching the code. On `main` it
also
gates the Docker publish, so the weekly base-image refresh cannot
republish
`:main` on a vulnerable dependency tree.

Policy lives in `.cargo/audit.toml` rather than in workflow flags, so a
bare
`cargo audit` locally gives the identical verdict. It runs `deny =
["warnings"]`
— unmaintained, unsound and yanked crates fail too, not just
vulnerabilities,
since letting those slide is precisely how #584 accumulated. The job
also runs
`cargo metadata --locked` first: `cargo audit` reads `Cargo.lock`, so a
lockfile
that disagrees with the manifests would make the gate meaningless.

`release.yml` is deliberately left alone. Its `release` job creates the
tag
*before* `build-container` runs, so a gate failing there would strand a
tagged
release with no artifacts. Releases are cut from `main`, which this gate
already
covers on every push and weekly.

`.github/dependabot.yml` covers the other half — weekly grouped PRs
(capped at
3 open) for direct crates and for GitHub Actions. Repo-level Dependabot
alerts
and security updates are now enabled too, which is what catches
transitive
crates between scheduled audit runs.

## Verification

- `cargo audit` — 0 vulnerabilities, 0 warnings, 563 crates.
- `cargo fmt --check`, `cargo clippy --all-targets -- -D warnings`, and
  `cargo test` (2064 passed, 0 failed) on the default workspace.
- `cargo clippy -p engine-zarr --features icechunk --all-targets -- -D
warnings`
and its `icechunk` test, so the feature-conditional graph is not hidden
by the
  default one (#584 acceptance criterion 4).
- `actionlint` on the workflow.

+++ crates/engine-grib/src/reader.rs
@@ -33,7 +33,7 @@ pub fn read_message(
                     entry.param
                 ))
             })?;
-            let file_size = meta.size as u64;
+            let file_size = meta.size;
             if file_size <= entry.offset {
                 return Err(DataServerError::Storage(format!(
                     "File {path} size {file_size} <= tail offset {}; \
